+++ totalsegmentator/bin/TotalSegmentator.py
@@ -23,11 +23,13 @@ def main():
                         type=lambda p: Path(p).absolute(), required=True)
 
     parser.add_argument("-o", metavar="directory", dest="output",
-                        help="Output directory for segmentation masks",
+                        help="Output directory for segmentation masks. " + 
+                             "Or path of multilabel output nifti file if --ml option is used." + 
+                             "Or path of output dicom seg file if --output_type is set to 'dicom_seg' or 'dicom_rtstruct'",
                         type=lambda p: Path(p).absolute(), required=True)
 
-    parser.add_argument("-ot", "--output_type", choices=["nifti", "dicom"],
-                    help="Select if segmentations shall be saved as Nifti or as Dicom RT Struct image.",
+    parser.add_argument("-ot", "--output_type", choices=["nifti", "dicom_rtstruct", "dicom_seg"],
+                    help="Select if segmentations shall be saved as Nifti, Dicom RT Struct, or Dicom SEG file.",
                     default="nifti")
 
     parser.add_argument("-ml", "--ml", action="store_true", help="Save one multilabel image for all classes",
@@ -63,7 +65,8 @@ def main():
                         "brain_structures", "liver_vessels", "oculomotor_muscles",
                         "thigh_shoulder_muscles", "thigh_shoulder_muscles_mr", "lung_nodules", "kidney_cysts", 
                         "breasts", "ventricle_parts", "aortic_sinuses", "liver_segments", "liver_segments_mr",
-                        "total_highres_test", "craniofacial_structures", "abdominal_muscles"],
+                        "total_highres_test", "craniofacial_structures", "abdominal_muscles", "teeth",
+                        "trunk_cavities"],
                         help="Select which model to use. This determines what is predicted.",
                         default="total")