+++ totalsegmentator/bin/totalseg_download_weights.py
@@ -63,6 +63,8 @@ def main():
         "liver_segments_mr": [576],
         "craniofacial_structures": [115],
         "abdominal_muscles": [952],
+        "teeth": [113],
+        "trunk_cavities": [343],
         
         "heartchambers_highres": [301],
         "appendicular_bones": [304],